Move another files into the unit test folder
